fix: hoist attributes builder
