+++ bin/client/src/l1/driver.rs
@@ -11,12 +11,13 @@ use alloy_primitives::B256;
 use anyhow::{anyhow, Result};
 use core::fmt::Debug;
 use kona_derive::{
+    attributes::StatefulAttributesBuilder,
     errors::PipelineErrorKind,
     pipeline::{DerivationPipeline, Pipeline, PipelineBuilder, StepResult},
     sources::EthereumDataSource,
     stages::{
         AttributesQueue, BatchQueue, ChannelBank, ChannelReader, FrameQueue, L1Retrieval,
-        L1Traversal, StatefulAttributesBuilder,
+        L1Traversal,
     },
     traits::{BlobProvider, ChainProvider, L2ChainProvider, OriginProvider},
 };
