+++ crates/derive/src/params.rs
@@ -1,10 +1,6 @@
 //! This module contains the parameters and identifying types for the derivation pipeline.
 
-use alloy_primitives::{address, b256, Address, B256};
-
-/// The sequencer fee vault address.
-pub const SEQUENCER_FEE_VAULT_ADDRESS: Address =
-    address!("4200000000000000000000000000000000000011");
+use alloy_primitives::{b256, B256};
 
 /// `keccak256("ConfigUpdate(uint256,uint8,bytes)")`
 pub const CONFIG_UPDATE_TOPIC: B256 =
